Field 54 in the ISO SPEC is 120 length `an LLL`
The fields beneath it are fixed 20 length additional amounts.

I've tried to write a composite field for it, but if the field is
present in the bitmap and zero length the library fails with:

```
failed to unpack message: failed to unpack field 54 (Additional
Amounts): failed to unpack subfield 01: not enough data to unpack,
expected: 20, got: 0
```

This patch "fixes" that, by just returning the offset if there's no
data in a composite spec.. But i'm not certain this is the correct
fix. (tests all pass, including these).